requirements.txt
torch
torchvision
torchaudio
numpy
pillow
scikit-image
pyyaml
tqdm
rich
accelerate
einops
matplotlib
